Error handling

+++ src/handlers/handler.js
@@ -1,4 +1,5 @@
 import prisma from "../db/db.js";
+import { validateAllFieldTypes } from "../validators/fieldValidators.js";
 
 const getAllStudents = async (req, res) => {
   try {
@@ -16,36 +17,60 @@ const getAllStudents = async (req, res) => {
     console.log(e);
     return res.json({
       message: "Failed to fetch students",
-      error: e,
+      stack: e?.message,
     });
   }
 };
 const getStudentById = async (req, res) => {
   try {
-    let matchStudent = await prisma.students.findUnique({
-      where: {
-        id: Number(req.params.id),
-      },
-    });
-    if (!matchStudent) {
-      return res.status(404).json({
-        message: "No student found with that id",
+    let { id } = req.params;
+    if (id == "") {
+      return res.status(400).json({
+        error: "Id cannot be empty",
       });
     }
+    if (isNaN(id)) {
+      return res.status(400).json({
+        error: "Id must be a number",
+      });
+    }
+    let matchStudent = await prisma.students.findUnique({
+      where: {
+        id: Number(id),
+      },
+    });
     res.status(200).json({
       message: "Student found",
       data: matchStudent,
     });
   } catch (e) {
+    if (e.code == "P2022") {
+      return res.status(404).json({
+        message: "No student found with that id",
+      });
+    }
     console.log(e);
     return res.json({
       message: "Failed to fetch student",
+      stack: e?.message,
     });
   }
 };
 const createStudent = async (req, res) => {
   try {
-    let data = req.body;
+    let { name, email } = req.body;
+    let validateMsg = validateAllFieldTypes("email", email);
+    if (validateMsg != null) {
+      return res.status(400).json({
+        error: validateMsg,
+      });
+    }
+    validateMsg = validateAllFieldTypes("name", name);
+    if (validateMsg != null) {
+      return res.status(400).json({
+        error: validateMsg,
+      });
+    }
     let createdStudent = await prisma.students.create({
       data: data,
     });
@@ -57,11 +82,23 @@ const createStudent = async (req, res) => {
     console.log(e);
     return res.json({
       message: "Faile to create student",
+      stack: e?.message,
     });
   }
 };
 const updateStudent = async (req, res) => {
   try {
+    let id = req.params;
+    if (id == "") {
+      return res.status(400).json({
+        error: "Id cannot be empty",
+      });
+    }
+    if (isNaN(id)) {
+      return res.status(400).json({
+        error: "Id must be a number",
+      });
+    }
     let data = req.body;
     let { name, email } = req.body;
     let updatedStudent = await prisma.students.update({
@@ -91,11 +128,23 @@ const updateStudent = async (req, res) => {
     console.log(e);
     return res.json({
       message: "Failed to update user",
+      stack: e?.message,
     });
   }
 };
 const deleteStudent = async (req, res) => {
   try {
+    let id = req.params;
+    if (id == "") {
+      return res.status(400).json({
+        error: "Id cannot be empty",
+      });
+    }
+    if (isNaN(id)) {
+      return res.status(400).json({
+        error: "Id must be a number",
+      });
+    }
     let deletedStudent = await prisma.students.delete({
       where: {
         id: Number(req.params.id),
@@ -114,6 +163,7 @@ const deleteStudent = async (req, res) => {
     console.log(e);
     return res.json({
       message: "Failed to delete student",
+      stack: e?.message,
     });
   }
 };
